+++ x-pack/plugins/ml/public/application/components/scatterplot_matrix/index.ts
@@ -4,5 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+export { useScatterplotFieldOptions } from './use_scatterplot_field_options';
 export { LEGEND_TYPES } from './scatterplot_matrix_vega_lite_spec';
 export { ScatterplotMatrix } from './scatterplot_matrix';
+export type { ScatterplotMatrixViewProps as ScatterplotMatrixProps } from './scatterplot_matrix_view';
