+++ x-pack/plugins/ml/public/application/components/scatterplot_matrix/scatterplot_matrix_vega_lite_spec.test.ts
@@ -163,6 +163,7 @@ describe('getScatterplotMatrixVegaLiteSpec()', () => {
       type: 'nominal',
     });
     expect(vegaLiteSpec.spec.encoding.tooltip).toEqual([
+      { field: 'the-color-field', type: 'nominal' },
       { field: 'x', type: 'quantitative' },
       { field: 'y', type: 'quantitative' },
     ]);
