+++ x-pack/plugins/ml/public/application/data_frame_analytics/common/use_results_view_config.ts
@@ -102,6 +102,12 @@ export const useResultsViewConfig = (jobId: string) => {
 
             try {
               indexP = await mlContext.indexPatterns.get(destIndexPatternId);
+
+              // Force refreshing the fields list here because a user directly coming
+              // from the job creation wizard might land on the page without the
+              // index pattern being fully initialized because it was created
+              // before the analytics job populated the destination index.
+              await mlContext.indexPatterns.refreshFields(indexP);
             } catch (e) {
               indexP = undefined;
             }
