+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_creation/components/configuration_step/configuration_step_form.tsx
@@ -27,10 +27,10 @@ import {
   TRAINING_PERCENT_MAX,
   FieldSelectionItem,
 } from '../../../../common/analytics';
+import { getScatterplotMatrixLegendType } from '../../../../common/get_scatterplot_matrix_legend_type';
 import { CreateAnalyticsStepProps } from '../../../analytics_management/hooks/use_create_analytics_form';
 import { Messages } from '../shared';
 import {
-  AnalyticsJobType,
   DEFAULT_MODEL_MEMORY_LIMIT,
   State,
 } from '../../../analytics_management/hooks/use_create_analytics_form/state';
@@ -51,18 +51,7 @@ import { SEARCH_QUERY_LANGUAGE } from '../../../../../../../common/constants/sea
 import { ExplorationQueryBarProps } from '../../../analytics_exploration/components/exploration_query_bar/exploration_query_bar';
 import { Query } from '../../../../../../../../../../src/plugins/data/common/query';
 
-import { LEGEND_TYPES, ScatterplotMatrix } from '../../../../../components/scatterplot_matrix';
-
-const getScatterplotMatrixLegendType = (jobType: AnalyticsJobType) => {
-  switch (jobType) {
-    case ANALYSIS_CONFIG_TYPE.CLASSIFICATION:
-      return LEGEND_TYPES.NOMINAL;
-    case ANALYSIS_CONFIG_TYPE.REGRESSION:
-      return LEGEND_TYPES.QUANTITATIVE;
-    default:
-      return undefined;
-  }
-};
+import { ScatterplotMatrix } from '../../../../../components/scatterplot_matrix';
 
 const requiredFieldsErrorText = i18n.translate(
   'xpack.ml.dataframe.analytics.createWizard.requiredFieldsErrorMessage',
@@ -498,6 +487,7 @@ export const ConfigurationStepForm: FC<CreateAnalyticsStepProps> = ({
                   : undefined
               }
               legendType={getScatterplotMatrixLegendType(jobType)}
+              searchQuery={jobConfigQuery}
             />
           </EuiPanel>
           <EuiSpacer />
