+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/expandable_section/expandable_section_splom.tsx
@@ -12,17 +12,14 @@ import { FormattedMessage } from '@kbn/i18n/react';
 
 import { EuiHorizontalRule, EuiSpacer } from '@elastic/eui';
 
-import { ScatterplotMatrix } from '../../../../../components/scatterplot_matrix';
+import {
+  ScatterplotMatrix,
+  ScatterplotMatrixProps,
+} from '../../../../../components/scatterplot_matrix';
 
 import { ExpandableSection } from './expandable_section';
 
-interface ExpandableSectionSplomProps {
-  fields: string[];
-  index: string;
-  resultsField?: string;
-}
-
-export const ExpandableSectionSplom: FC<ExpandableSectionSplomProps> = (props) => {
+export const ExpandableSectionSplom: FC<ScatterplotMatrixProps> = (props) => {
   const splomSectionHeaderItems = undefined;
   const splomSectionContent = (
     <>
