+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/exploration_page_wrapper/exploration_page_wrapper.tsx
@@ -9,16 +9,21 @@ import React, { FC, useCallback, useState } from 'react';
 import { EuiCallOut, EuiFlexGroup, EuiFlexItem, EuiPanel, EuiSpacer, EuiText } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 
+import { getAnalysisType, getDependentVar } from '../../../../../../../common/util/analytics_utils';
+
+import { useScatterplotFieldOptions } from '../../../../../components/scatterplot_matrix';
+
 import {
   defaultSearchQuery,
+  getScatterplotMatrixLegendType,
   useResultsViewConfig,
   DataFrameAnalyticsConfig,
 } from '../../../../common';
-import { ResultsSearchQuery } from '../../../../common/analytics';
+import { ResultsSearchQuery, ANALYSIS_CONFIG_TYPE } from '../../../../common/analytics';
 
 import { DataFrameTaskStateType } from '../../../analytics_management/components/analytics_list/common';
 
-import { ExpandableSectionAnalytics } from '../expandable_section';
+import { ExpandableSectionAnalytics, ExpandableSectionSplom } from '../expandable_section';
 import { ExplorationResultsTable } from '../exploration_results_table';
 import { ExplorationQueryBar } from '../exploration_query_bar';
 import { JobConfigErrorCallout } from '../job_config_error_callout';
@@ -99,6 +104,14 @@ export const ExplorationPageWrapper: FC<Props> = ({
     language: pageUrlState.queryLanguage,
   };
 
+  const resultsField = jobConfig?.dest.results_field ?? '';
+  const scatterplotFieldOptions = useScatterplotFieldOptions(
+    indexPattern,
+    jobConfig?.analyzed_fields.includes,
+    jobConfig?.analyzed_fields.excludes,
+    resultsField
+  );
+
   if (indexPatternErrorMessage !== undefined) {
     return (
       <EuiPanel grow={false}>
@@ -124,6 +137,9 @@ export const ExplorationPageWrapper: FC<Props> = ({
       />
     );
   }
+
+  const jobType =
+    jobConfig && jobConfig.analysis ? getAnalysisType(jobConfig?.analysis) : undefined;
 
   return (
     <>
@@ -178,6 +194,27 @@ export const ExplorationPageWrapper: FC<Props> = ({
       {isLoadingJobConfig === false && jobConfig !== undefined && isInitialized === true && (
         <EvaluatePanel jobConfig={jobConfig} jobStatus={jobStatus} searchQuery={searchQuery} />
       )}
+
+      {isLoadingJobConfig === true && jobConfig === undefined && <LoadingPanel />}
+      {isLoadingJobConfig === false &&
+        jobConfig !== undefined &&
+        isInitialized === true &&
+        typeof jobConfig?.id === 'string' &&
+        scatterplotFieldOptions.length > 1 &&
+        typeof jobConfig?.analysis !== 'undefined' && (
+          <ExpandableSectionSplom
+            fields={scatterplotFieldOptions}
+            index={jobConfig?.dest.index}
+            color={
+              jobType === ANALYSIS_CONFIG_TYPE.REGRESSION ||
+              jobType === ANALYSIS_CONFIG_TYPE.CLASSIFICATION
+                ? getDependentVar(jobConfig.analysis)
+                : undefined
+            }
+            legendType={getScatterplotMatrixLegendType(jobType)}
+            searchQuery={searchQuery}
+          />
+        )}
 
       {isLoadingJobConfig === true && jobConfig === undefined && <LoadingPanel />}
       {isLoadingJobConfig === false &&
