+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -14232,8 +14232,8 @@
     "xpack.ml.splom.dynamicSizeLabel": "動的サイズ",
     "xpack.ml.splom.fieldSelectionLabel": "フィールド",
     "xpack.ml.splom.fieldSelectionPlaceholder": "フィールドを選択",
-    "xpack.ml.splom.RandomScoringLabel": "ランダムスコアリング",
-    "xpack.ml.splom.SampleSizeLabel": "サンプルサイズ",
+    "xpack.ml.splom.randomScoringLabel": "ランダムスコアリング",
+    "xpack.ml.splom.sampleSizeLabel": "サンプルサイズ",
     "xpack.ml.splom.toggleOff": "オフ",
     "xpack.ml.splom.toggleOn": "オン",
     "xpack.ml.splomSpec.outlierScoreThresholdName": "異常スコアしきい値：",
