+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -14271,8 +14271,8 @@
     "xpack.ml.splom.dynamicSizeLabel": "动态大小",
     "xpack.ml.splom.fieldSelectionLabel": "字段",
     "xpack.ml.splom.fieldSelectionPlaceholder": "选择字段",
-    "xpack.ml.splom.RandomScoringLabel": "随机评分",
-    "xpack.ml.splom.SampleSizeLabel": "样例大小",
+    "xpack.ml.splom.randomScoringLabel": "随机评分",
+    "xpack.ml.splom.sampleSizeLabel": "样例大小",
     "xpack.ml.splom.toggleOff": "关闭",
     "xpack.ml.splom.toggleOn": "开启",
     "xpack.ml.splomSpec.outlierScoreThresholdName": "离群值分数阈值：",
