+++ x-pack/test/functional/services/ml/index.ts
@@ -17,6 +17,7 @@ import { MachineLearningDataFrameAnalyticsProvider } from './data_frame_analytic
 import { MachineLearningDataFrameAnalyticsCreationProvider } from './data_frame_analytics_creation';
 import { MachineLearningDataFrameAnalyticsEditProvider } from './data_frame_analytics_edit';
 import { MachineLearningDataFrameAnalyticsResultsProvider } from './data_frame_analytics_results';
+import { MachineLearningDataFrameAnalyticsScatterplotProvider } from './data_frame_analytics_scatterplot';
 import { MachineLearningDataFrameAnalyticsMapProvider } from './data_frame_analytics_map';
 import { MachineLearningDataFrameAnalyticsTableProvider } from './data_frame_analytics_table';
 import { MachineLearningDataVisualizerProvider } from './data_visualizer';
@@ -63,6 +64,9 @@ export function MachineLearningProvider(context: FtrProviderContext) {
   const dataFrameAnalyticsResults = MachineLearningDataFrameAnalyticsResultsProvider(context);
   const dataFrameAnalyticsMap = MachineLearningDataFrameAnalyticsMapProvider(context);
   const dataFrameAnalyticsTable = MachineLearningDataFrameAnalyticsTableProvider(context);
+  const dataFrameAnalyticsScatterplot = MachineLearningDataFrameAnalyticsScatterplotProvider(
+    context
+  );
 
   const dataVisualizer = MachineLearningDataVisualizerProvider(context);
   const dataVisualizerTable = MachineLearningDataVisualizerTableProvider(context, commonUI);
@@ -105,6 +109,7 @@ export function MachineLearningProvider(context: FtrProviderContext) {
     dataFrameAnalyticsResults,
     dataFrameAnalyticsMap,
     dataFrameAnalyticsTable,
+    dataFrameAnalyticsScatterplot,
     dataVisualizer,
     dataVisualizerFileBased,
     dataVisualizerIndexBased,
